requirements.txt
fastapi
uvicorn
passlib
pydantic
starlette
mongoengine
jinja2
python-multipart
python-dotenv
pymongo[srv]

